SW-68 Split Components
