+++ .gitignore
@@ -35,3 +35,6 @@ yarn-error.log*
 .env
 .env.development
 .env.production
+
+# package-lock
+package-lock.json
