+++ app/diary/[diaryId]/page.jsx
@@ -5,6 +5,7 @@ import diaryData from '../../../public/data/dairy.json';
 export default function diaryShow() {
 
   const diary = diaryData;
+  
   return (
     <>
       <div className="px-10 py-12 lg:px-28">
@@ -15,7 +16,7 @@ export default function diaryShow() {
             <div className="text-2xl font-extrabold">{diary.date}</div>
             <div className="flex flex-wrap mt-3">
                 {diary.keywords.map((keyword, idx) => (
-                    <div key={keyword} className="px-3 mb-2 py-1 mr-2.5 text-sm font-medium text-slate-500 bg-slate-200 rounded-xl dark:bg-slate-200 dark:text-slate-800 ">
+                    <div key={keyword} className="px-3 mb-2 py-1 mr-2.5 text-sm font-medium text-zinc-500 bg-zinc-200 rounded-xl dark:bg-zinc-200 dark:text-zinc-800 ">
                         #{keyword.keyword}
                     </div>
                 ))}
@@ -41,7 +42,7 @@ export default function diaryShow() {
             <div className="text-xl">
               <button className="mx-2 text-base font-semibold duration-200 btn btn-secondary hover:scale-105">수정하기</button>
               <button className="mx-2 text-base font-semibold duration-200 btn btn-accent hover:scale-105">삭제하기</button>
-              <Link href="/diary/list/grid" className="mx-2 text-base font-semibold duration-200 border-opacity-0 outline-none text-slate-50 bg-slate-400 hover:bg-slate-500 btn outline-0 border-spacing-0 hover:scale-105">목록으로</Link>
+              <Link href="/diary/list/grid" className="mx-2 text-base font-semibold duration-200 border-opacity-0 outline-none text-zinc-50 bg-zinc-400 hover:bg-zinc-500 btn outline-0 border-spacing-0 hover:scale-105">목록으로</Link>
             </div>
           </div>
         </div>
