+++ app/diary/common/backButton.js
@@ -8,7 +8,7 @@ export default function BackButton() {
       <div className="flex w-full m-3 justify-items-start">
           <div>
             <Link href="/diary/list/grid">
-                <ChevronLeftIcon className="block w-6 h-6 text-slate-600"/>
+                <ChevronLeftIcon className="block w-6 h-6 text-zinc-600"/>
             </Link>
           </div>
       </div>
