+++ app/diary/common/diaryInputFormat.js
@@ -6,7 +6,13 @@ import dynamic from 'next/dynamic';
 
 export default function DiaryInputFormat(props) {  
   const Editor = dynamic(() => import('../../common/Editor'), { ssr: false });
+
+  useEffect(() => {
+    setEditorLoaded(true);
+  }, []);
+  
   const [editorLoaded, setEditorLoaded] = useState(false);
+
   const resultView = useRef(null);
   
   const onClick = (str) => {
@@ -14,9 +20,6 @@ export default function DiaryInputFormat(props) {
       resultView.current.innerHTML = `<h2>html결과 view입니다</h2>${str}`;
     }
   };
-  useEffect(() => {
-    setEditorLoaded(true);
-  }, []);
 
   return (
     <>
@@ -29,12 +32,14 @@ export default function DiaryInputFormat(props) {
               onClick={onClick}
               editorLoaded={editorLoaded}
               value={props.Contents}
+              date={props.date}
             />
           :
           <Editor
             name="description"
             onClick={onClick}
             editorLoaded={editorLoaded}
+            date={props.date}
           />
         }
       </div>
