+++ app/diary/dashboard/page.jsx
@@ -23,24 +23,6 @@ export default function diaryPage() {
                         <CardSocialTraffic />
                     </div>
                 </div>
-                {/* <div className="w-full grid-cols-6">
-                    <div className="flex col-span-6">
-                        <div className="col-span-6 px-4 mb-12 sm:col-span-4 xl:mb-0">
-                        <CardLineChart />
-                        </div>
-                        <div className="col-span-6 px-4 sm:col-span-2">
-                        <CardBarChart />
-                        </div>
-                    </div>
-                    <div className="flex flex-wrap col-span-6 mt-4">
-                        <div className="px-4 mb-12 xl:w-8/12 xl:mb-0">
-                        <CardPageVisits />
-                        </div>
-                        <div className="w-full px-4 xl:w-4/12">
-                        <CardSocialTraffic />
-                        </div>
-                    </div>
-                </div> */}
             </div>
         </>
     )
