+++ app/home/page.jsx
@@ -1,10 +1,6 @@
 export default function homePage() {
     return ( 
         <>
-            <div className="bg-blue-50">
-                대시보드 페이지<br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/><br/>
-                dslfdn
-            </div>
         </>
     )
 }
